Split ExternalCommandGateway into intake helpers (351 → 163 lines)
The class kept four concerns inline: normalization heuristics, the
orchestration-preview builder, file IO for the command log, and
the actual receive() orchestration. Each now lives in its own file
under ./external-command-gateway/:

- types.ts      shared shapes (ExternalCommandInput, NormalizedCommand,
                GatewayResult, CommandLogFile), COMMAND_LOG_PATH,
                SUPPORTED_CHANNELS, ENTITY_PATTERNS with a comment
                explaining the first-match-wins ordering
- normalize.ts  message → NormalizedCommand: extractEntities,
                summarizeIntent, estimateConfidence (jsdoc explains
                the per-channel score deltas), estimateApproval
                (conservative gating policy)
- preview.ts    buildOrchestrationPreview — always non-executing,
                comment explains why both branches pass
                `approved: false` (intake is not allowed to execute
                as a side effect; only /api/operational/orchestrate
                can, gated by caller-supplied flag)
- log.ts        logCommand + listRecentCommands — append-only JSON
                with disk-failure-warned-not-thrown semantics

The class itself keeps static `receive`, `normalize`, `listRecent`
so all 5 internal callers (VoiceCommandBridge, MessagingBridge,
registerIntakeRoutes) keep working unchanged.
